Created TerrainLayerAsset
