+++ editor/src/main/com/mbrlabs/mundus/editor/ui/modules/dialogs/AddTerrainChunksDialog.kt
@@ -316,8 +316,9 @@ class AddTerrainChunksDialog : BaseDialog("Add Terrain Chunks") {
                     val chessboard =
                         projectManager.current().assetManager.findAssetByID(EditorAssetManager.STANDARD_ASSET_TEXTURE_CHESSBOARD)
                     if (chessboard != null) {
-                        asset.splatBase = chessboard as TextureAsset
-                        asset.applyDependencies()
+                        val terrainLayerAsset = asset.terrainLayerAsset
+                        terrainLayerAsset.splatBase = chessboard as TextureAsset
+                        terrainLayerAsset.applyDependencies()
                         metaSaver.save(asset.meta)
                     }
